+++ components/ui/shell.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react';
+
+export function Shell({ children }: { children: ReactNode }) {
+  return <div className="mx-auto w-full max-w-7xl">{children}</div>;
+}
